Feature/1 updated autonomous code

See merge request team1746/1746_rapidreact_2022!23

+++ src/main/java/frc/robot/Constants.java
@@ -5,15 +5,12 @@
 package frc.robot;
 
 /**
- * The Constants class provides a convenient place for teams to hold robot-wide
- * numerical or boolean
- * constants. This class should not be used for any other purpose. All constants
- * should be declared
+ * The Constants class provides a convenient place for teams to hold robot-wide numerical or boolean
+ * constants. This class should not be used for any other purpose. All constants should be declared
  * globally (i.e. public static). Do not put anything functional in this class.
  *
  * <p>
- * It is advised to statically import this class (or one of its inner classes)
- * wherever the
+ * It is advised to statically import this class (or one of its inner classes) wherever the
  * constants are needed, to reduce verbosity.
  */
 public final class Constants {
@@ -65,7 +62,7 @@ public final class Constants {
         public static final int kShooterSlave2 = 43;
         public static final int kShooterSlave3 = 42;
         public static final int kIndexerMotor = 30;
-        public static final int kFullPower = 1;
+        public static final double kFullPower = 1;
         public static final int kZeroPower = 0;
 
     }
